+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release 2020.4 (?)
 
+## New features
+- Added merging of similar compile requests to the compile queue (#2137)
+
 # Release 2020.3 (2020-07-02)
 
 ## New features